+++ fixtures/backup/model_dependencies/sorted.json
@@ -21,12 +21,14 @@
   "sentry.DeletedProject",
   "sentry.DeletedTeam",
   "sentry.Email",
+  "sentry.Environment",
   "sentry.EventAttachment",
   "sentry.EventUser",
   "sentry.GroupCommitResolution",
   "sentry.GroupRedirect",
   "sentry.GroupRelease",
   "sentry.Organization",
+  "sentry.Release",
   "sentry.IdentityProvider",
   "sentry.DocIntegration",
   "sentry.ExternalActor",
@@ -40,15 +42,29 @@
   "sentry.OrganizationMapping",
   "sentry.OrganizationMemberMapping",
   "sentry.OrgAuthToken",
+  "sentry.SnubaQuery",
+  "sentry.SnubaQueryEventType",
+  "sentry.Project",
+  "sentry.ProjectBookmark",
+  "sentry.ProjectKey",
+  "sentry.ProjectOwnership",
   "sentry.ProjectPlatform",
+  "sentry.ProjectRedirect",
   "sentry.PromptsActivity",
   "sentry.PullRequest",
   "sentry.PullRequestComment",
+  "sentry.RawEvent",
   "sentry.RecentSearch",
   "sentry.RelayUsage",
   "sentry.Relay",
+  "sentry.ReleaseActivity",
+  "sentry.ReleaseEnvironment",
   "sentry.ReleaseFile",
+  "sentry.ReleaseProjectEnvironment",
   "sentry.Repository",
+  "sentry.ReprocessingReport",
+  "sentry.Rule",
+  "sentry.RuleActivity",
   "sentry.SavedSearch",
   "sentry.ScheduledDeletion",
   "sentry.RegionScheduledDeletion",
@@ -56,15 +72,24 @@
   "sentry.ServiceHook",
   "sentry.RegionTombstone",
   "sentry.ControlTombstone",
+  "sentry.ProjectTransactionThresholdOverride",
+  "sentry.ProjectTransactionThreshold",
   "sentry.UserEmail",
   "sentry.UserIP",
   "sentry.UserPermission",
   "sentry.UserReport",
   "sentry.UserRole",
   "sentry.UserRoleUser",
+  "sentry.NotificationAction",
   "sentry.TimeSeriesSnapshot",
+  "sentry.AlertRule",
+  "sentry.AlertRuleTrigger",
+  "sentry.AlertRuleTriggerAction",
+  "sentry.AlertRuleActivity",
+  "sentry.DiscoverSavedQuery",
   "sentry.Monitor",
   "sentry.MonitorLocation",
+  "sentry.MonitorEnvironment",
   "sentry.MetricsKeyIndexer",
   "sentry.StringIndexer",
   "sentry.PerfStringIndexer",
@@ -73,17 +98,44 @@
   "nodestore.Node",
   "replays.ReplayRecordingSegment",
   "social_auth.UserSocialAuth",
+  "sentry.MonitorCheckIn",
+  "sentry.DiscoverSavedQueryProject",
+  "sentry.AlertRuleExcludedProjects",
+  "sentry.Incident",
+  "sentry.IncidentSeen",
+  "sentry.IncidentProject",
+  "sentry.NotificationActionProject",
   "sentry.ServiceHookProject",
+  "sentry.RuleSnooze",
+  "sentry.QuerySubscription",
+  "sentry.ProcessingIssue",
+  "sentry.OrganizationOnboardingTask",
   "sentry.LostPasswordHash",
+  "sentry.LatestAppConnectBuildsCheck",
+  "sentry.RepositoryProjectPathConfig",
+  "sentry.ProjectIntegration",
   "sentry.OrganizationIntegration",
   "sentry.Identity",
+  "sentry.GroupTombstone",
+  "sentry.ReleaseProject",
+  "sentry.OrganizationMember",
   "sentry.Team",
+  "sentry.OrganizationMemberTeam",
+  "sentry.Group",
+  "sentry.GroupHistory",
   "sentry.FeatureAdoption",
+  "sentry.EnvironmentProject",
+  "sentry.Distribution",
+  "sentry.Deploy",
   "sentry.DashboardTombstone",
+  "sentry.Dashboard",
+  "sentry.DashboardProject",
+  "sentry.Counter",
   "sentry.Commit",
   "sentry.BroadcastSeen",
   "sentry.UserAvatar",
   "sentry.TeamAvatar",
+  "sentry.ProjectAvatar",
   "sentry.OrganizationAvatar",
   "sentry.DocIntegrationAvatar",
   "sentry.FileBlobIndex",
@@ -96,9 +148,12 @@
   "sentry.AssistantActivity",
   "sentry.ArtifactBundleFlatFileIndex",
   "sentry.ArtifactBundle",
+  "sentry.AppConnectBuild",
   "sentry.ApiApplication",
   "sentry.UserOption",
+  "sentry.ProjectOption",
   "sentry.OrganizationOption",
+  "sentry.Activity",
   "sentry.ApiAuthorization",
   "sentry.ApiGrant",
   "sentry.ApiToken",
@@ -108,54 +163,6 @@
   "sentry.DebugIdArtifactBundle",
   "sentry.ProjectArtifactBundle",
   "sentry.CommitFileChange",
-  "sentry.OrganizationMember",
-  "sentry.SentryApp",
-  "sentry.PagerDutyService",
-  "sentry.SentryAppComponent",
-  "sentry.SentryAppInstallation",
-  "sentry.SentryAppInstallationForProvider",
-  "sentry.SentryAppInstallationToken",
-  "sentry.OrganizationAccessRequest",
-  "sentry.Project",
-  "sentry.ProjectBookmark",
-  "sentry.ProjectKey",
-  "sentry.ProjectOwnership",
-  "sentry.ProjectRedirect",
-  "sentry.PullRequestCommit",
-  "sentry.RawEvent",
-  "sentry.ReprocessingReport",
-  "sentry.Rule",
-  "sentry.RuleActivity",
-  "sentry.ProjectTransactionThresholdOverride",
-  "sentry.ProjectTransactionThreshold",
-  "sentry.NotificationAction",
-  "sentry.DiscoverSavedQuery",
-  "sentry.DiscoverSavedQueryProject",
-  "sentry.NotificationActionProject",
-  "sentry.ProjectTeam",
-  "sentry.ProcessingIssue",
-  "sentry.OrganizationOnboardingTask",
-  "sentry.LatestAppConnectBuildsCheck",
-  "sentry.RepositoryProjectPathConfig",
-  "sentry.ProjectIntegration",
-  "sentry.GroupTombstone",
-  "sentry.Release",
-  "sentry.ReleaseProject",
-  "sentry.OrganizationMemberTeam",
-  "sentry.Group",
-  "sentry.GroupHistory",
-  "sentry.Environment",
-  "sentry.EnvironmentProject",
-  "sentry.Distribution",
-  "sentry.Deploy",
-  "sentry.Dashboard",
-  "sentry.DashboardProject",
-  "sentry.Counter",
-  "sentry.SentryAppAvatar",
-  "sentry.ProjectAvatar",
-  "sentry.AppConnectBuild",
-  "sentry.ProjectOption",
-  "sentry.Activity",
   "sentry.DashboardWidget",
   "sentry.GroupOwner",
   "sentry.GroupAssignee",
@@ -172,35 +179,28 @@
   "sentry.GroupShare",
   "sentry.GroupSnooze",
   "sentry.GroupSubscription",
+  "sentry.SentryApp",
+  "sentry.PagerDutyService",
+  "sentry.SentryAppComponent",
+  "sentry.SentryAppInstallation",
+  "sentry.SentryAppInstallationForProvider",
+  "sentry.SentryAppInstallationToken",
+  "sentry.OrganizationAccessRequest",
   "sentry.PlatformExternalIssue",
   "sentry.EventProcessingIssue",
-  "sentry.SnubaQuery",
-  "sentry.SnubaQueryEventType",
-  "sentry.QuerySubscription",
+  "sentry.ProjectTeam",
   "sentry.ProjectCodeOwners",
-  "sentry.ReleaseActivity",
+  "sentry.PullRequestCommit",
   "sentry.ReleaseCommit",
-  "sentry.ReleaseEnvironment",
   "sentry.ReleaseHeadCommit",
-  "sentry.ReleaseProjectEnvironment",
   "sentry.RuleFireHistory",
-  "sentry.AlertRule",
-  "sentry.AlertRuleActivity",
-  "sentry.TeamKeyTransaction",
-  "sentry.MonitorEnvironment",
-  "sentry.MonitorCheckIn",
-  "sentry.AlertRuleExcludedProjects",
-  "sentry.Incident",
-  "sentry.IncidentSeen",
-  "sentry.IncidentProject",
-  "sentry.RuleSnooze",
-  "sentry.DashboardWidgetQuery",
   "sentry.PendingIncidentSnapshot",
   "sentry.IncidentSnapshot",
   "sentry.IncidentActivity",
   "sentry.IncidentSubscription",
-  "sentry.AlertRuleTrigger",
+  "sentry.IncidentTrigger",
   "sentry.AlertRuleTriggerExclusion",
-  "sentry.AlertRuleTriggerAction",
-  "sentry.IncidentTrigger"
+  "sentry.TeamKeyTransaction",
+  "sentry.DashboardWidgetQuery",
+  "sentry.SentryAppAvatar"
 ]
